게시판 기초 구현

+++ src/main/java/com/estsoft/repository/BoardRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.estsoft.domain.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Long> {} 
+public interface BoardRepository extends JpaRepository<Board, Integer> {}
